feat: add native parsing for xsd:float
Implement native parsing support for xsd:float.
Refactor the parsing logic to use a switch-case structure for better
readability and maintainability, replacing the previous if-else chain.
Integer and double cases are now handled separately, allowing the use of
the most suitable standard library parsing methods. This change also
eliminates the use of fmt.Sprintf for float-to-integer checks, improving
performance.

Signed-off-by: Michael Adler <michael.adler@siemens.com>